🚀 Create separate middleware for user storage
