+++ server/src/routes/userRoutes.js
@@ -2,10 +2,7 @@ const express = require("express");
 const userController = require("../controllers/userController");
 const ensureAuthenticated = require("../middlewares/ensureAuthenticated");
 const authorize = require("../middlewares/authorize");
-const {
-  uploadUserImg,
-  resizeUserImg,
-} = require("../controllers/userController");
+const { uploadUserImg, resizeUserImg } = require("../middlewares/userStorage");
 
 const router = express.Router();
